+++ numba_dpex/core/parfors/reduction_helper.py
@@ -17,10 +17,10 @@ from numba.core.ir_utils import (
 from numba.parfors import parfor
 from numba.parfors.parfor_lowering_utils import ParforLoweringBuilder
 
-from numba_dpex import utils
 from numba_dpex.core.datamodel.models import (
     dpex_data_model_manager as kernel_dmm,
 )
+from numba_dpex.core.utils.cgutils_extra import get_llvm_type
 from numba_dpex.core.utils.kernel_launcher import KernelLaunchIRBuilder
 from numba_dpex.dpctl_iface import libsyclinterface_bindings as sycl
 
@@ -431,11 +431,11 @@ class ReductionKernelVariables:
 
             dest = builder.bitcast(
                 lowerer.getvar(redvar),
-                utils.get_llvm_type(context=context, type=types.voidptr),
+                get_llvm_type(context=context, type=types.voidptr),
             )
             src = builder.bitcast(
                 builder.load(array_attr),
-                utils.get_llvm_type(context=context, type=types.voidptr),
+                get_llvm_type(context=context, type=types.voidptr),
             )
 
             args = [
